// labserveroriginal/web/admin/src/app/models/entity/student.ts
import { IEntityBase } from "./entity-base";
import { IGroup } from "./group";

export interface IStudent extends IEntityBase {
    groupId: number;
    name: string;
    email: string;
    username: string | undefined;
    gitLabUserId: number | undefined;
    initialPassword: string | undefined;

    group: IGroup | undefined;
}